+++ Assets/Spells/SpellDescriptor.cs
@@ -7,19 +7,24 @@ public class SpellDescriptor
 {
     public string Name { get; private set; }
 
-    Dictionary<string, float> opinions;
+    Dictionary<string, double> opinions;
 
-    public SpellDescriptor(string name, Dictionary<string, float> opinions)
+    public SpellDescriptor(string name, Dictionary<string, double> opinions)
     {
         this.Name = name;
         this.opinions = opinions;
     }
 
-    public float GetOpinion(Region region)
+    public double GetOpinion(Region region)
     {
         if (!opinions.ContainsKey(region.InternalName))
             throw new ArgumentException("No opinion data available for region " + region.InternalName + " (" + region.DisplayedName + ")");
 
         return opinions[region.InternalName];
     }
+
+    //public string GetOpinionStrings()
+    //{
+    //    return string.Join(" ", opinions.Values.Select(u => u.ToString("f0")).ToArray());
+    //}
 }
